Changed obviously positive integer types to size_t.

+++ src/configurations/RBTConnectConfig.cpp
@@ -4,11 +4,11 @@
 
 #include "RBTConnectConfig.h"
 
-unsigned long RBTConnectConfig::MAX_NUM_ITER    = 1e9;
-unsigned long RBTConnectConfig::MAX_NUM_STATES  = 1e9;
-float RBTConnectConfig::MAX_PLANNING_TIME       = 60;
-float RBTConnectConfig::D_CRIT                  = 0.03;
-float RBTConnectConfig::DELTA                   = 3.14159;
-int RBTConnectConfig::NUM_SPINES                = 7;
-int RBTConnectConfig::NUM_ITER_SPINE            = 5;
-bool RBTConnectConfig::USE_EXPANDED_BUBBLE      = true;
+size_t RBTConnectConfig::MAX_NUM_ITER       = 1e9;
+size_t RBTConnectConfig::MAX_NUM_STATES     = 1e9;
+float RBTConnectConfig::MAX_PLANNING_TIME   = 60;
+float RBTConnectConfig::D_CRIT              = 0.03;
+float RBTConnectConfig::DELTA               = 3.14159;
+size_t RBTConnectConfig::NUM_SPINES         = 7;
+size_t RBTConnectConfig::NUM_ITER_SPINE     = 5;
+bool RBTConnectConfig::USE_EXPANDED_BUBBLE  = true;
